feat: create a related prompt component
